requirements.txt
APScheduler==3.9.1
pymongo==4.1.0
python-decouple==3.6
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
six==1.16.0
tzdata==2022.1
tzlocal==4.2
